requirements.txt
ib_insync>=0.9.86
# alpaca-trade-api>=3.0.0  # kept as optional — only needed if re-enabling Alpaca crypto
pandas>=2.0
numpy>=1.24
ta>=0.11.0
pyyaml>=6.0
python-dotenv>=1.0
flask>=3.0
lightgbm>=4.0
scikit-learn>=1.3
discord.py>=2.0
vaderSentiment>=3.3
requests>=2.31
